Split OLED processing into separate files.
